attempting to fix map

+++ OnTheMap/Model/Responses/Student.swift
@@ -11,8 +11,8 @@ import Foundation
 class Student: Codable {
     var firstName: String!
     var lastName: String!
-    var longitude: Float!
-    var latitude: Float!
+    var longitude: Double!
+    var latitude: Double!
     var mapString: String!
     var mediaURL: String!
     var uniqueKey: String!
